Add support for running CrustalRuby without Crystal in dry mode
Introduce a new config option `crystal_missing_ignore`, which defaults to `false`.
When enabled, the application can run without the `crystal` binary.
If `crystal` is missing, the library will not be compiled,
but the Ruby application will still run as-is.

This change simplifies onboarding by removing the requirement for all developers to install `crystal`.